tests: move to lib universe, observatory and  supercomputer.
